[bcr-wdc-webapi] Move random identity data to test utils and export
