+++ crates/bcr-wdc-webapi/src/lib.rs
@@ -6,5 +6,6 @@ pub mod quotes;
 pub mod signatures;
 pub mod swap;
 pub mod wallet;
+pub mod test_utils;
 
 // ----- end imports
